Split editor out into site handlers and formatters
A handler detects if it should be applied to a site based on DOM or domain. It then contains special code that makes it fit better on that page.

Additionally it can use a formatter to load and save in whatever format is required for that page - markdown, html, Trac wiki etc

Support is added for:
- Trac
- Github
- WP editor pages (for example, bbPress)
- Plain textareas